requirements.txt
django==4.2.10
djangorestframework==3.14.0
drf-spectacular==0.27.1
python-dotenv==1.0.1
psycopg2-binary
black==24.2.0